refactor: [#1298] move HttpApiContainer to tracker-api-core package
