+++ packages/tracker-api-core/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod container;
 pub mod statistics;
