+++ src/container.rs
@@ -17,6 +17,7 @@ use bittorrent_udp_tracker_core::container::UdpTrackerCoreContainer;
 use bittorrent_udp_tracker_core::services::banning::BanService;
 use bittorrent_udp_tracker_core::{self, MAX_CONNECTION_ID_ERRORS_PER_IP};
 use tokio::sync::RwLock;
+use torrust_tracker_api_core::container::HttpApiContainer;
 use torrust_tracker_configuration::{Configuration, Core, HttpApi, HttpTracker, UdpTracker};
 use tracing::instrument;
 
@@ -91,17 +92,6 @@ impl AppContainer {
     }
 }
 
-pub struct HttpApiContainer {
-    pub core_config: Arc<Core>,
-    pub http_api_config: Arc<HttpApi>,
-    pub in_memory_torrent_repository: Arc<InMemoryTorrentRepository>,
-    pub keys_handler: Arc<KeysHandler>,
-    pub whitelist_manager: Arc<WhitelistManager>,
-    pub ban_service: Arc<RwLock<BanService>>,
-    pub http_stats_repository: Arc<bittorrent_http_tracker_core::statistics::repository::Repository>,
-    pub udp_stats_repository: Arc<bittorrent_udp_tracker_core::statistics::repository::Repository>,
-}
-
 /// It initializes the IoC Container.
 #[instrument(skip())]
 pub fn initialize_app_container(configuration: &Configuration) -> AppContainer {
@@ -142,32 +132,4 @@ pub fn initialize_app_container(configuration: &Configuration) -> AppContainer {
         http_stats_repository,
         udp_stats_repository,
     }
-}
-
-#[must_use]
-pub fn initialize_http_api_container(core_config: &Arc<Core>, http_api_config: &Arc<HttpApi>) -> Arc<HttpApiContainer> {
-    let tracker_core_container = TrackerCoreContainer::initialize(core_config);
-
-    // HTTP stats
-    let (_http_stats_event_sender, http_stats_repository) =
-        bittorrent_http_tracker_core::statistics::setup::factory(core_config.tracker_usage_statistics);
-    let http_stats_repository = Arc::new(http_stats_repository);
-
-    // UDP stats
-    let (_udp_stats_event_sender, udp_stats_repository) =
-        bittorrent_udp_tracker_core::statistics::setup::factory(core_config.tracker_usage_statistics);
-    let udp_stats_repository = Arc::new(udp_stats_repository);
-
-    let ban_service = Arc::new(RwLock::new(BanService::new(MAX_CONNECTION_ID_ERRORS_PER_IP)));
-
-    Arc::new(HttpApiContainer {
-        http_api_config: http_api_config.clone(),
-        core_config: core_config.clone(),
-        in_memory_torrent_repository: tracker_core_container.in_memory_torrent_repository.clone(),
-        keys_handler: tracker_core_container.keys_handler.clone(),
-        whitelist_manager: tracker_core_container.whitelist_manager.clone(),
-        ban_service: ban_service.clone(),
-        http_stats_repository: http_stats_repository.clone(),
-        udp_stats_repository: udp_stats_repository.clone(),
-    })
 }
