+++ src/servers/apis/routes.rs
@@ -16,6 +16,7 @@ use axum::routing::get;
 use axum::{middleware, BoxError, Router};
 use hyper::{Request, StatusCode};
 use torrust_server_lib::logging::Latency;
+use torrust_tracker_api_core::container::HttpApiContainer;
 use torrust_tracker_configuration::{AccessTokens, DEFAULT_TIMEOUT};
 use tower::timeout::TimeoutLayer;
 use tower::ServiceBuilder;
@@ -30,7 +31,6 @@ use tracing::{instrument, Level, Span};
 use super::v1;
 use super::v1::context::health_check::handlers::health_check_handler;
 use super::v1::middlewares::auth::State;
-use crate::container::HttpApiContainer;
 use crate::servers::apis::API_LOG_TARGET;
 
 /// Add all API routes to the router.
