+++ src/servers/apis/v1/routes.rs
@@ -2,9 +2,9 @@
 use std::sync::Arc;
 
 use axum::Router;
+use torrust_tracker_api_core::container::HttpApiContainer;
 
 use super::context::{auth_key, stats, torrent, whitelist};
-use crate::container::HttpApiContainer;
 
 /// Add the routes for the v1 API.
 pub fn add(prefix: &str, router: Router, http_api_container: &Arc<HttpApiContainer>) -> Router {
